Add kuri_api reference package
This is a fork of robot_api that is specific to Kuri. While some
parts of robot_api are easy to make generic, or don't cause conflicts, the increasing use
of Mayfield types and APIs means there is greater and greater
nuisance to Fetch users. We'll track stubs and reference implementations
in this folder from now on